% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf_expression.R
\name{bnf_expression}
\alias{bnf_expression}
\title{bnf_expression expression from BNF spec}
\usage{
bnf_expression(
  spec,
  global_spec,
  zero_more_lambda = 1,
  one_more_lambda = 1,
  zero_prob = 0.5,
  state = list(N = 1L)
)
}
\arguments{
\item{spec}{A BNF spec element in list form}

\item{global_spec}{A global BNF spec in list form}

\item{zero_more_lambda}{poisson lambda for distribution of zero or mores}

\item{one_more_lambda}{poisson lambda for distribution of one or mores}

\item{zero_prob}{probability of a zero in for zero or one case}

\item{state}{any state information to carry around.}
}
\value{
code generated from the given BNF
}
\description{
bnf_expression expression from BNF spec
}
