% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf_expression.R
\name{bnf_function}
\alias{bnf_function}
\title{Make a function from a string}
\usage{
bnf_function(expr)
}
\arguments{
\item{expr}{A string (output from \code{bnf_expression})}
}
\value{
A function
}
\description{
Make a function from a string
}
