% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{char_expression_to_function}
\alias{char_expression_to_function}
\title{Make a function from a character string representing an R expression}
\usage{
char_expression_to_function(expr)
}
\arguments{
\item{expr}{A string representing an R expression e.g. \code{x + y}}
}
\value{
A function with 2 formal arguments (\code{x}, \code{y}) which will
calculate the value of the given expression at this location
}
\description{
Make a function from a character string representing an R expression
}
\examples{
\dontrun{
f <- char_expression_to_function("x + y")
f(1, 2)
}
}
