% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf_data_generation.R
\name{dat_generation_grid}
\alias{dat_generation_grid}
\title{Generate xy data from an expression}
\usage{
dat_generation_grid(f, xdim = 20, xby = 1, ydim = 20, yby = 1)
}
\arguments{
\item{f}{A function that accepts two numeric arguments and returns a single numeric value}

\item{xdim}{An integer defining length of the x dimension}

\item{xby, yby}{seq steps in x and y directions}

\item{ydim}{An integer defining length of the y dimension}
}
\value{
A tibble
}
\description{
Generate xy data from an expression
}
