% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf_data_generation.R
\name{dat_generation_path}
\alias{dat_generation_path}
\title{Generate path data from an expression}
\usage{
dat_generation_path(fun, seed1, seed2, n)
}
\arguments{
\item{fun}{function}

\item{seed1}{A double}

\item{seed2}{A double}

\item{n}{number of points}
}
\value{
A tibble
}
\description{
Generate path data from an expression
}
