% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_grid}
\alias{eval_grid}
\title{Evaluate code (given as a character string) at gridded points on the (x, y) plane}
\usage{
eval_grid(code, xmin = 0, xmax = 1, xn = 10, ymin = xmin, ymax = xmax, yn = xn)
}
\arguments{
\item{code}{A character string containing an R expression.}

\item{xmin, xmax, xn}{grid dimensions along x-axis i.e. the minimum and maximum
extents on the x axis, and the number of grid points to generate
along this dimension}

\item{ymin, ymax, yn}{grid dimensions along y-axis i.e. the minimum and maximum
extents on the y axis, and the number of grid points to generate
along this dimension. Defaults to same as x-axis}
}
\value{
data.frame of \code{(x, y)} coordinates and the evaluated function at these
coordinates \code{z}
}
\description{
Evaluate code (given as a character string) at gridded points on the (x, y) plane
}
\examples{
eval_grid("x + y")

}
