% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_bnf}
\alias{parse_bnf}
\title{Parse BNF text into an R data structure}
\usage{
parse_bnf(bnf_text)
}
\arguments{
\item{bnf_text}{string containing the BNF rules}
}
\value{
nested list representing the BNF
}
\description{
TODO: detail what flavour of BNF is supported.
}
