% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_bnf_text_to_tokens}
\alias{parse_bnf_text_to_tokens}
\title{Parse BNF text into tokens}
\usage{
parse_bnf_text_to_tokens(bnf_text)
}
\arguments{
\item{bnf_text}{BNF text}
}
\value{
named character vector of tokens
}
\description{
Parse BNF text into tokens
}
