% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_bnf_tokens_to_spec}
\alias{parse_bnf_tokens_to_spec}
\title{Parse tokens lexed from BNF text into the R nested list representation}
\usage{
parse_bnf_tokens_to_spec(tokens)
}
\arguments{
\item{tokens}{named character vector of tokens from lex()}
}
\value{
R nested list representation of the BNF grammar
}
\description{
Parse tokens lexed from BNF text into the R nested list representation
}
