% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_items}
\alias{parse_items}
\title{Parse items}
\usage{
parse_items(tokens)
}
\arguments{
\item{tokens}{named character vector of tokens}
}
\value{
a list representation of the items
}
\description{
Parse items
}
