% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf_parse_grammar_to_list.R
\name{parse_bnf_tokens_to_spec}
\alias{parse_bnf_tokens_to_spec}
\title{Parse tokens lexed from BNF}
\usage{
parse_bnf_tokens_to_spec(tokens)
}
\arguments{
\item{tokens}{named character vector of tokens from lex()}
}
\value{
Internal BNF representation
}
\description{
Parse tokens lexed from BNF
}
